Add files via upload
This iteration of the Challenge 6 file is updated with an aggregation of the San Francisco housing data between the years of 2010 and 2016. The data is grouped by year and an average of the total number of housing units within each year is plotted and compared using the hvplot function. Through the data I found that  the number of housing units over the 6 year period increases each year, trending positively.